move file utilities to new module
